Add Tabbed UI, Bounds Preview mode, and Float support for movement delay
- power.py: delay_between_movements now supports floats
- laser.py: run loop respects Preview Mode (power=3) to pause random movement
- app.py: added /api/preview endpoint, updated settings to support floats
- index.html: implemented Bootstrap tabs, vertical bounds list with Preview Point buttons, and auto-stop preview on tab exit

+++ src/power.py
@@ -20,7 +20,7 @@ class ControlPower:
 
     # Movement setters
     def set_delay_between_movements(self, delay):
-        self.delay_between_movements = int(delay)
+        self.delay_between_movements = float(delay)
 
     def set_percentage_move_chance(self, speed):
         self.percentage_move_chance = float(speed)
